Commit: Changed authorizations and changed put method

+++ src/main/java/com/example/demo/domain/group/GroupServiceImpl.java
@@ -2,16 +2,19 @@ package com.example.demo.domain.group;
 
 import com.example.demo.domain.appUser.User;
 import com.example.demo.domain.appUser.UserRepository;
-import com.example.demo.domain.appUser.UserService;
 import com.example.demo.domain.role.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import javax.management.InstanceAlreadyExistsException;
+import javax.management.InstanceNotFoundException;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -35,31 +38,20 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public void delete(UUID uuid) {
-        groupRepository.delete(findById(uuid));
-    }
-
-    /*@Override
-    public Group put(Group newGroup, UUID uuid) throws InstanceNotFoundException {
-        Optional<Group> foundGroup = groupRepository.findById(uuid);
-        if (foundGroup.isPresent()){
-            foundGroup.get().setName(newGroup.getName());
-            foundGroup.get().setDescription(newGroup.getDescription());
-            foundGroup.get().setUsers(newGroup.getUsers());
-
-        }else {
-            throw new InstanceNotFoundException("");
+    public String delete(UUID uuid) throws InstanceNotFoundException {
+        if (groupRepository.existsById(uuid)){
+            groupRepository.delete(findById(uuid));
+            return "Group deleted";
+        } else {
+            throw new InstanceNotFoundException("Group not found");
         }
-
-        return groupRepository.save(foundGroup.get());
-    }*/
+    }
 
     @Override
     public Group saveGroup(Group group) throws InstanceAlreadyExistsException {
-        if (groupRepository.findByName(group.getName()) != null){
+        if (groupRepository.findByName(group.getName()) != null) {
             throw new InstanceAlreadyExistsException("Group already exists");
-        }
-        else {
+        } else {
             return groupRepository.save(group);
         }
     }
@@ -86,30 +78,50 @@ public class GroupServiceImpl implements GroupService {
     public void addUserToGroup(String username, String groupName) {
         User user = userRepository.findByUsername(username);
         Group group = groupRepository.findByName(groupName);
-        if (user.getGroup() == null){
+        if (user.getGroup() == null) {
             group.getUsers().add(user);
             groupRepository.save(group);
-        }else {
+        } else {
             System.out.println("Already in group");
         }
     }
 
     @Override
     public Group findById(UUID id) {
-            Optional<Group> group = groupRepository.findById(id);
-            return group.orElse(null);
+        Optional<Group> group = groupRepository.findById(id);
+        return group.orElse(null);
     }
 
     @Override
     public Group updateGroup(UUID id, Group group) {
         return groupRepository.findById(id)
                 .map(group1 -> {
+                    groupRepository.deleteById(id);
                     group1.setName(group.getName());
                     group1.setDescription(group.getDescription());
                     group1.setUsers(group.getUsers());
-                    return groupRepository.save(group);
+                    return groupRepository.save(group1);
                 }).orElseGet(() -> {
+                    group.setId(id);
                     return groupRepository.save(group);
                 });
+    }
+
+    @Override
+    public boolean isUserAuthorizedForGroup(UUID uuid) {
+        Object auth = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        User currentUser = userRepository.findByUsername(((UserDetails)auth).getUsername());
+
+        for (Role role : currentUser.getRoles()) {
+            if (role.getName().equals("ADMIN")){
+                return true;
+            }
+        }
+
+        if (currentUser.getGroup().getId().equals(uuid)){
+            return true;
+        }else {
+            return false;
+        }
     }
 }
